LayoutTests/imported/w3c/web-platform-tests/css/css-values/if-supports-quirks-expected.txt

test

FAIL CSS Values and Units Test: CSS inline if() function supports() in quirks mode Error: assert_equals: expected "true_value" but got "if(supports(width: 30): true_value; else: false_value;)"
